a few changes, mainly quest, userException.jsp, sendEmail.jsp and controller
