+++ src/main/java/com/containers/controller/ContainerController.java
@@ -1,9 +1,11 @@
 package com.containers.controller;
 
+import com.containers.model.Container;
 import com.containers.service.ContainerService;
+import org.dom4j.rule.Mode;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
@@ -17,65 +19,38 @@ public class ContainerController {
         this.containerService = containerService;
     }
 
-    @GetMapping("/")
-    public ModelAndView getHomePage(){
-        return new ModelAndView("index");
-    }
-
-    @GetMapping("/login")
-    public ModelAndView getLoginPage() {
-        return new ModelAndView("login");
-    }
-
-    @GetMapping("/guest")
-    public ModelAndView getGuestPage() {
-        return new ModelAndView("guest");
-    }
-
-    @PostMapping("/menu")
-    public ModelAndView getMenuPage() {
-        return new ModelAndView("menu");
-    }
-
-    @GetMapping("/container/add")
-    public ModelAndView addContainerPage() {
+    @GetMapping("/containers/add")
+    public ModelAndView containerAdd() {
         return new ModelAndView("addNewContainer");
     }
 
-    @GetMapping("/damages/history")
-    public ModelAndView getDamagesHistory() {
-        return new ModelAndView("damagesHistory");
+    @GetMapping("/containers/add/model")
+    public ModelAndView containersAddModel() {
+        return new ModelAndView("containerModel");
+    }
+
+    @GetMapping("/containers/find")
+    public ModelAndView findContainer(@RequestParam String containerNo) {
+        ModelAndView modelAndView = new ModelAndView("findContainer");
+        modelAndView.addObject("findContainer", containerService.findContainerById(containerNo));
+        return modelAndView;
+    }
+
+    @PostMapping("/containers/update")
+    public ModelAndView updateContainer(@ModelAttribute Container container) {
+        ModelAndView modelAndView = new ModelAndView("updateContainer");
+        modelAndView.addObject("updateContainer", containerService.updateContainer(container));
+        return modelAndView;
+    }
+
+    @GetMapping("/containers/delete")
+    public String deleteContainer(@RequestParam String containerNo) {
+        containerService.deleteContainer(containerNo);
+        return "redirect:/containers";
     }
 
     @GetMapping("/container/status")
     public ModelAndView getDamageStatus() {
         return new ModelAndView("containerStatus");
-    }
-
-    @GetMapping("/email/send")
-    public ModelAndView sendEmail() {
-        return new ModelAndView("sendEmail");
-    }
-
-    @GetMapping("/users/search")
-    public ModelAndView searchUsers() {
-        return new ModelAndView("searchUsers");
-    }
-
-    @GetMapping("user/edit")
-    public ModelAndView editUser(){
-        return new ModelAndView("editUser");
-    }
-
-    @GetMapping("/container/add/model")
-    public ModelAndView getContainerModelPage() {
-        return new ModelAndView("containerModel");
-    }
-
-    @GetMapping("/container/find")
-    public ModelAndView findContainer(@RequestParam String containerNo) {
-        ModelAndView modelAndView = new ModelAndView("findContainer");
-        modelAndView.addObject("findContainer", containerService.findContainerById(containerNo));
-        return modelAndView;
     }
 }
