+++ src/main/webapp/WEB-INF/views/userException.jsp
@@ -4,7 +4,73 @@
 
 <!DOCTYPE html>
 <html lang="en">
+<head>
+    <title>Exception</title>
+    <style>
+        * {
+            box-sizing: border-box;
+        }
+
+        body, html {
+            margin: 0;
+        }
+
+        body {
+            font-family: Arial, serif;
+            font-size: 17px;
+        }
+
+        .outside {
+            position: relative;
+        }
+
+        .outside img {
+            background-position: center;
+            background-repeat: no-repeat;
+            background-size: cover;
+        }
+
+        .inside {
+            position: absolute;
+            width: 100%;
+        }
+
+    </style>
+</head>
 <body>
-<h1>${message}</h1>
+<div class="outside">
+
+    <img src="/images/bg_con.jpg" alt="container" title="containerHome" style="
+        height: 100%;
+        width: 100%;">
+
+    <div class="inside"
+         style="position: absolute;
+         top: 10%;
+         background-size: cover;
+         -webkit-background-clip: text;
+         padding-right: 20%;">
+
+        <div class="background" style="position: relative;
+        background-size: cover;
+        left: 50%;
+        border-radius: 5px;
+        width: 250px;
+        height: 250px;">
+
+            <div class="form" style="position: absolute;
+            top: 20%;
+            left: 30%;">
+                <table>
+                    <h1>${message}</h1>
+                </table>
+
+                <button onclick="window.location.href='/'" type="submit" class="pure-button pure-button-primary">
+                    Back to main page
+                </button>
+            </div>
+        </div>
+    </div>
+</div>
 </body>
 </html>
